Multiple devices/hub support
Makes it possible to connect multiple devices using a USB hub.

Currently the mappings are global, not per-device, so for example if
you connect two mice, the left buttons will be mapped to the same
thing on both of them.

+++ firmware/src/pio_usb_stuff.h
@@ -34,7 +34,7 @@ extern "C" {
 #include "pio_usb.h"
 }
 
-typedef void (*report_handler_t)(uint8_t* report, int len, uint8_t interface);
+typedef void (*report_handler_t)(uint8_t* report, int len, uint16_t interface);
 
 void launch_pio_usb();
 void pio_usb_task(report_handler_t);
